+++ FelicaLib_Remodeled/UnitTest45/FelicaUtilityTest.cs
@@ -36,24 +36,17 @@ namespace UnitTest45
         }
 
         [TestMethod]
-        public void GetEdyBalance_1()
+        public void ReadWithoutEncryption_1()
         {
-            var target = FelicaHelper.GetEdyBalance();
-            Assert.AreEqual(12345, target);
+            var target = FelicaUtility.ReadWithoutEncryption(0xFE00, 0x1317, 0);
+            Assert.AreEqual(12345, target.ToEdyBalance());
         }
 
         [TestMethod]
-        public void GetWaonBalance_1()
+        public void ReadWithoutEncryption_2()
         {
-            var target = FelicaHelper.GetWaonBalance();
-            Assert.AreEqual(12345, target);
-        }
-
-        [TestMethod]
-        public void GetSuicaBalance_1()
-        {
-            var target = FelicaHelper.GetSuicaBalance();
-            Assert.AreEqual(12345, target);
+            var target = FelicaUtility.ReadWithoutEncryption(FelicaSystemCode.Edy, 0x1317, 0);
+            Assert.AreEqual(12345, target.ToEdyBalance());
         }
     }
 }
